+++ src/utils/json.test.ts
@@ -2,7 +2,7 @@ import { readFileSync } from 'node:fs'
 import path from 'node:path'
 import { describe, expect, test } from 'vitest'
 import { writeFixtures } from '../../tests/utils.ts'
-import { writeJsonFile } from './json.ts'
+import { detectIndentation, writeJsonFile } from './json.ts'
 
 describe('writeJsonFile', () => {
   test('creates a new file when it does not exist', async (context) => {
@@ -73,3 +73,26 @@ describe('writeJsonFile', () => {
     expect(readFileSync(filePath, 'utf8')).toBe('{\n  "foo": "baz"\n}\n')
   })
 })
+
+describe('detectIndent', () => {
+  test('two spaces', ({ expect }) => {
+    expect(detectIndentation(stringifyJson(2))).toBe(2)
+  })
+  test('four spaces', ({ expect }) => {
+    expect(detectIndentation(stringifyJson(4))).toBe(4)
+  })
+  test('tab', ({ expect }) => {
+    expect(detectIndentation(stringifyJson('\t'))).toBe('\t')
+  })
+  test('empty', ({ expect }) => {
+    expect(detectIndentation('')).toBe(2)
+  })
+  test('empty line', ({ expect }) => {
+    expect(detectIndentation('{\n\n  "foo": 42 }')).toBe(2)
+  })
+})
+
+function stringifyJson(indentation: string | number): string {
+  const contents = JSON.stringify({ foo: 42 }, null, indentation)
+  return contents
+}
